🎨 Move random_circuit.py to the graphix dir
