+++ tests/test_transpiler.py
@@ -6,7 +6,7 @@ from numpy.random import PCG64, Generator
 
 import graphix.pauli
 import graphix.simulator
-import tests.random_circuit as rc
+import graphix.random_circuit as rc
 from graphix.transpiler import Circuit
